Moved API call to its own file and connected it to Repository
